+++ app/src/androidTest/java/nextstep/signup/PasswordValidationTest.kt
@@ -3,7 +3,7 @@ package nextstep.signup
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.ui.screen.PasswordTextField
+import nextstep.signup.ui.component.textfield.PasswordTextField
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
