+++ app/src/main/java/nextstep/signup/ui/component/textfield/PasswordTextField.kt
@@ -1,0 +1,33 @@
+package nextstep.signup.ui.component.textfield
+
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
+import nextstep.signup.R
+import nextstep.signup.ui.validatePassword
+
+@Composable
+fun PasswordTextField(
+    password: String,
+    onPasswordChange: (String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val validationState = remember { derivedStateOf { validatePassword(password) } }
+
+    SignupTextField(
+        value = password,
+        onValueChange = onPasswordChange,
+        label = stringResource(R.string.password),
+        keyboardOptions = KeyboardOptions(
+            autoCorrectEnabled = false,
+            keyboardType = KeyboardType.Password
+        ),
+        secure = true,
+        validationState = validationState.value,
+        modifier = modifier
+    )
+}
