+++ app/src/main/java/nextstep/signup/ui/screen/SignupScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
@@ -17,19 +16,16 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import nextstep.signup.R
-import nextstep.signup.ui.ValidationState
 import nextstep.signup.ui.component.SignupButton
-import nextstep.signup.ui.component.SignupTextField
+import nextstep.signup.ui.component.textfield.EmailTextField
+import nextstep.signup.ui.component.textfield.PasswordConfirmTextField
+import nextstep.signup.ui.component.textfield.PasswordTextField
+import nextstep.signup.ui.component.textfield.UsernameTextField
 import nextstep.signup.ui.theme.SignupTheme
-import nextstep.signup.ui.validateEmail
-import nextstep.signup.ui.validatePassword
-import nextstep.signup.ui.validatePasswordConfirm
-import nextstep.signup.ui.validateUsername
 
 @Composable
 fun SignupScreen(
@@ -52,7 +48,12 @@ fun SignupScreen(
             ),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        TitleContent()
+        Text(
+            text = stringResource(R.string.welcome_to_compose),
+            fontWeight = FontWeight.W700,
+            fontSize = 26.sp,
+            lineHeight = 20.sp,
+        )
         Spacer(modifier = Modifier.height(42.dp))
         TextFieldsContent(
             username = username,
@@ -76,16 +77,6 @@ fun SignupScreen(
 }
 
 @Composable
-private fun TitleContent() {
-    Text(
-        text = stringResource(R.string.welcome_to_compose),
-        fontWeight = FontWeight.W700,
-        fontSize = 26.sp,
-        lineHeight = 20.sp,
-    )
-}
-
-@Composable
 private fun TextFieldsContent(
     username: String,
     email: String,
@@ -102,122 +93,27 @@ private fun TextFieldsContent(
         verticalArrangement = Arrangement.spacedBy(33.dp)
     ) {
         UsernameTextField(
+            modifier = Modifier.fillMaxWidth(),
             username = username,
             onUsernameChange = onUsernameChange
         )
         EmailTextField(
+            modifier = Modifier.fillMaxWidth(),
             email = email,
             onEmailChange = onEmailChange
         )
         PasswordTextField(
+            modifier = Modifier.fillMaxWidth(),
             password = password,
             onPasswordChange = onPasswordChange
         )
         PasswordConfirmTextField(
+            modifier = Modifier.fillMaxWidth(),
             password = password,
             passwordConfirm = passwordConfirm,
             onPasswordConfirmChange = onPasswordConfirmChange
         )
     }
-}
-
-@Composable
-fun UsernameTextField(
-    username: String,
-    onUsernameChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val errorMsg = validateUsername(username)
-
-    SignupTextField(
-        value = username,
-        onValueChange = onUsernameChange,
-        label = stringResource(R.string.username),
-        errorMsg = if (errorMsg is ValidationState.Error) {
-            stringResource(errorMsg.resourceId)
-        } else {
-            null
-        },
-        modifier = modifier.fillMaxWidth()
-    )
-}
-
-@Composable
-fun EmailTextField(
-    email: String,
-    onEmailChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val errorMsg = validateEmail(email)
-
-    SignupTextField(
-        value = email,
-        onValueChange = onEmailChange,
-        label = stringResource(R.string.email),
-        keyboardOptions = KeyboardOptions(
-            autoCorrectEnabled = false,
-            keyboardType = KeyboardType.Email
-        ),
-        errorMsg = if (errorMsg is ValidationState.Error) {
-            stringResource(errorMsg.resourceId)
-        } else {
-            null
-        },
-        modifier = modifier.fillMaxWidth()
-    )
-}
-
-@Composable
-fun PasswordTextField(
-    password: String,
-    onPasswordChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val errorMsg = validatePassword(password)
-
-    SignupTextField(
-        value = password,
-        onValueChange = onPasswordChange,
-        label = stringResource(R.string.password),
-        keyboardOptions = KeyboardOptions(
-            autoCorrectEnabled = false,
-            keyboardType = KeyboardType.Password
-        ),
-        secure = true,
-        errorMsg = if (errorMsg is ValidationState.Error) {
-            stringResource(errorMsg.resourceId)
-        } else {
-            null
-        },
-        modifier = modifier.fillMaxWidth()
-    )
-}
-
-@Composable
-fun PasswordConfirmTextField(
-    password: String,
-    passwordConfirm: String,
-    onPasswordConfirmChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val errorMsg = validatePasswordConfirm(password, passwordConfirm)
-
-    SignupTextField(
-        value = passwordConfirm,
-        onValueChange = onPasswordConfirmChange,
-        label = stringResource(R.string.password_confirm),
-        keyboardOptions = KeyboardOptions(
-            autoCorrectEnabled = false,
-            keyboardType = KeyboardType.Password
-        ),
-        secure = true,
-        errorMsg = if (errorMsg is ValidationState.Error) {
-            stringResource(errorMsg.resourceId)
-        } else {
-            null
-        },
-        modifier = modifier.fillMaxWidth()
-    )
 }
 
 @Preview(
